+++ mayan/apps/documents/models/__init__.py
@@ -5,3 +5,4 @@ from .document_version_models import *  # NOQA
 from .document_version_page_models import *  # NOQA
 from .document_type_models import *  # NOQA
 from .misc_models import *  # NOQA
+from .recently_accessed_document_models import *  # NOQA